Add connectQueryParent which provides abstracted functionality to connectQuery subclasses.
